Credit Card Payments with Stripe
Created Stripe free account to test tour purchase orders. Installed stripe and added bookingController and bookingRoutes to define the payment handler and the route for the checkout session. Mounted bookingRouter in app.js. Added a new request for bookings in postman.

Integrated Stripe into the back-end after setting up the booking route(s) and booking controller for the purpose of handling payments with the API. Setup a stripe.js file with a bookTour function to export/import into index.js to handle payment processing from the front-end.

Created a bookings schema/model to implement with the booking controller. Temporary booking checkout method (workaround) used upon completing a booking payment to create a booking document. Temporary because the method is unsecure and only for development; method will be changed specifically for production using Stripe Webhooks.

Added a getMyTours function which will read and display all the tours that are booked by the currently logged in user. Completed the Bookings API after implementing in the CRUD operations/methods from the handlerFactory file within the Bookings controller and routes.